# backend/swagger/v1/swagger.yaml
openapi: 3.0.1
info:
  title: GroceryStore API V1
  version: v1
  description: API for GroceryStore application with Products, Orders, and Users
paths:
...
  /api/v1/products:
    get:
      summary: List all products
      tags:
        - Products
      responses:
        '200':
          description: successful
    post:
      summary: Create a product
      tags:
        - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product:
                  type: object
                  properties:
                    name:
                      type: string
                    price:
                      type: number
                    stock:
                      type: integer
                    description:
                      type: string
                    category:
                      type: string
                    brand:
                      type: string
      responses:
        '201':
          description: product created
  /api/v1/products/{id}:
    get:
      summary: Get a product
      tags:
        - Products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful
    patch:
      summary: Update a product
      tags:
        - Products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful
    delete:
      summary: Delete a product
      tags:
        - Products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: no content
...
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
  - url: http://localhost:5001
    description: Development server